Migrate over to using `enabledApiProposals` in `package.json`
