+++ package.json
@@ -18,7 +18,7 @@
     },
     "languageServerVersion": "0.5.30",
     "publisher": "ms-python",
-    "enableProposedApi": true,
+    "enabledApiProposals": ["languageStatus", "quickPickSortByLabel", "quickPickSeparators", "testRefresh", "testObserver", "notebookEditor"],
     "author": {
         "name": "Microsoft Corporation"
     },
@@ -1954,7 +1954,7 @@
         "@types/tmp": "0.0.33",
         "@types/untildify": "^3.0.0",
         "@types/uuid": "^3.4.3",
-        "@types/vscode": "~1.53.0",
+        "@types/vscode": "~1.63.0",
         "@types/winreg": "^1.2.30",
         "@types/xml2js": "^0.4.2",
         "@typescript-eslint/eslint-plugin": "^3.7.0",
