+++ src/client/common/application/types.ts
@@ -522,7 +522,7 @@ export interface IDocumentManager {
     /**
      * The currently visible editors or an empty array.
      */
-    readonly visibleTextEditors: TextEditor[];
+    readonly visibleTextEditors: readonly TextEditor[];
 
     /**
      * An [event](#Event) which fires when the [active editor](#window.activeTextEditor)
@@ -542,7 +542,7 @@ export interface IDocumentManager {
      * An [event](#Event) which fires when the array of [visible editors](#window.visibleTextEditors)
      * has changed.
      */
-    readonly onDidChangeVisibleTextEditors: Event<TextEditor[]>;
+    readonly onDidChangeVisibleTextEditors: Event<readonly TextEditor[]>;
 
     /**
      * An [event](#Event) which fires when the selection in an editor has changed.
@@ -874,7 +874,7 @@ export interface IDebugService {
     /**
      * List of breakpoints.
      */
-    readonly breakpoints: Breakpoint[];
+    readonly breakpoints: readonly Breakpoint[];
 
     /**
      * An [event](#Event) which fires when the [active debug session](#debug.activeDebugSession)
